Embed and hardcode vertex shader sources in drawable.cpp
Embed and hardcode vertex shader sources in drawable.cpp. Similar
to the built-in material shaders the vertex shaders for the built-in
drawable shapes can be considered built-in and harcoded directly
within the c++ source code. This removes any need to do any runtime
data deployment for the graphics library to function when using its
primitive/built-in functionality only.

This change also allows 3 other things/hacks to be removed:

- When packing workspace resources there's no longer need to copy any
  glsl files directly since any shader file is now referenced by
  a custom material only and would get packed when that material is
  packed.

- No need to have any hacks for diverting unresolved shader paths
  when playing the game in GSEditor into editor's installation path.
  Any unresolved raw resource path is now only ever coming from the
  application itself (if using a custom engine) and can always be
  mapped relative to the game processes working directory.

- No need to have to cmake install rules to copy the shader sources
  around all over the place. Hooray \o/