Fix round rect drawable corner distortion on non-square rects
Fix round rect drawable corner distortion on non-square rects. The
problem is that when the view matrix is applied on the vertices of
the square round rect the shape may no longer be square when
rasterized. The difference between the X and Y axis will result in
distorted corners. In order to fix this we must look at the provided
view matrix and try to figure out the result of the view matrix on
the extents of the model and then compute new sizes for the corner
boxes so that they always remain square on screen when rasterized.

In order to limit the number of geometries to some maximum that will
be generated the potential range of aspect ratios is clamped to a
few selected ratios. This a tradeoff between memory, performance
and correct rendering.