Fix capsule drawable end point distortion on non-square transforms
Fix capsule drawable end point distortion on non-square transforms
The problem is that when the view matrix is applied on the vertices
of the capsule the shape may no longer be within a square box when
rasterized. The difference between the X and Y axis will result in
distorted end caps. In order to fix this we must look at the provided
view matrix and try to figure out the result of the view matrix on
the extents of the model and then compute new sizes for the end caps
so that they always remain within a square on screen when rasterized.

In order to limit the number of geometries to some maximum that will
be generated the potential range of aspect ratios is clamped to a
few selected ratios. This a tradeoff between memory, performance
and correct rendering. However unlike similar functionality regarding
the rounded rect corners the capsule has some noticeable jumps
when selecting between possible geometries and aspect ratios. The code
to select for the closest geometry allows for half steps but is still
not 100% smooth. This might require further fine tuning.