Use a fixed round rect class id based in workspace settings
Use a fixed round rect class id based in workspace settings. Since
the class settings are hard coded (corner radius) and this is then
one of the "primitive" drawables configured in the editor the class
id needs also be fixed to some particular id in order to let this
shape be easily referred by it's "known id".