Add several new audio features and some new audio elements
Add several new audio features and some new audio elements. The new
elements include:
- StereoMaker to turn a mono audio stream into stereo stream by
  copying the mono channel into one of the stereo channels.
- Fader element for creating audio fade in/out effect by interpolating
  audio gain value (between 0.0f and 1.0f) over time.
- ZeroSource provides endless audio buffers filled with zeros.
- SourceMixer combines (mixes) several input sources into a single
  output stream. Unlike Mixer it doesn't offer any input port
  but aggregates the elements directly.

Command and event dispatching. Command mechanism is a way to send
commands via the audio player to a source object playing in the
audio thread. The command can alter the destination source in some
way. For example when the source object is an AudioGraph the gain
value of a Gain element in the graph can be modified.

Event mechanism is a way for the source object in the player's
audio thread dispatch back some status etc events to notify the
client of the audio system of some interesting events such as the
completion of some source.