Add first version of audio engine/subsystem in the game lib
Add first version of audio engine/subsystem in the game lib. The
engine provides an API to manipulate two audio playback streams.
One stream is the so called "music" stream and the other one is
the so called "effects or FX" stream. The client (game) can send
commands to manipulate either audio graph and both streams can
support an arbitrary number of mixer sources.

There are still couple of action items to implement:

- Proper loading of audio data using some loader interface. This
  at minimum requires some kind of mapping of the resource URIs
  to a packaged resource location.

- Get rid of the zero source. Currently the audio device and source
  system doesn't support implicit pause/resume when the audio source
  doesn't produce any audio buffers. This needs to be implemented in
  the audio system then zero source can be removed in order to reduce
  the CPU and memory allocator load.