Handle oversized audio buffer coming from an audio graph
Handle oversized audio buffer coming from an audio graph. In case
the audio graph produces and audio buffer that exceeds the current
capacity of the playback stream write buffer store the remaining
buffer contents into an intermediate buffer and write into the
playback stream's buffer on subsequent fill.

This buffering size mismatch can happen whenever a buffer computed
based on of the stream size on the Nth fill buffer call is delivered
on the 1+Nth call.