Add audio player source/stream progress reporting through a message
Add audio player source/stream progress reporting through a message.
This provides a way to get some information about the current state
of the stream in terms of playback time and bytes sent to the
device. The information is provided through a message because of the
complexity of threading. Also the message is only provided on request
in order to reduce the overhead of generating needless messages.