Add audio widget refresh request on timer when playing audio
Add audio widget refresh request on timre when playing audio. The
problem is that the Update function cannot be used since the audio
widget is not an accelearated widget so that'd basically run a busy
loop. However the widget's refresh UI low frequency update function
is sluggish since it only triggers approx at every 500ms.