Copy original audio buffer tags over when creating a new audio buffer
Copy original audio buffer tags over when creating a new audio buffer
as the outgoing buffer in a non-source audio element. This is needed
in order to not lose any buffer tag information that exists and has
been accumulated over element traversal.

Inspect all buffer tags for buffer source completion in Playlist
before determining whether to move onto the next source port or not.
The input buffer could be the result of mixing for example and thus
playlist cannot consider this source port to be done until all
buffer sources are done.