Disable library name suffix on debug build
Disable library name suffix on debug build. This should make it
easier to switch between debug and release build since the way the
engine library is loaded doesn't need to change to account for a
change in the library name.

Add the DEBUG string to the application title on a debug build to
make it obvious which version is being used.