Refactor the concept of "debug pause" into game engine from host
Refactor the concept of "debug pause" into game engine from host.
The host doesn't know and cannot possibly implement everything that
might need to be done in order to support "pausing the game" fully.
For example pausing the current game audio playback as well as not
advancing the animation/physics.