Refactor GfxWidget VSYNC implementation and surface creation
Refactor GfxWidget VSYNC implementation and surface creation. Try
to simplify the logic regarding when to create a surface with
swap interval enabled and when to do a synced swap. Add a silly
thread wait when synced swap is not used, i.e. when VSYNC is off
or when the vsync surface is not exposed (according to Qt/libANGLE
on Windows this is undefined behaviour). The wait is used in order
to try to throttle the execution in order to spare some cycles and
battery when running on a laptop.

There's now a new BUG though that on current Linux installation the
vsync doesn't actually happen. A second bug is that toggling the
VSYNC (which causes the surface recreation renders black until
swapping the tab to another tab and back in the MainWindow).