Synchronize structural changes to entity to any open animation widget
Synchronize structural changes to entity to any open animation widget.
The entity widget and the animation widget are tightly coupled and
share the same underlying entity representation. If the entity's node
hierarchy is changed the animation track widget must be notified
in order to make sure that there are no stale data on the animation UI
for node IDs etc.