Fix potential missing call to swapBuffers when using a dialog
Fix potential missing call to swapBuffers when using a dialog. The
commit 2429d07 restructured the GfxWidget logic regarding the
buffer swapping. However it also introduced a bug where any dialog
that would use a GfxWidget would not swap because any such dialog
(as of now) does not put the application in the accelerated rendering
loop. This means that there'd be no call to GfxWindow::EndFrame and
thus no call to swapBuffers. Moving the call to swapBuffers back
into the paintGL function. The fight goes on!