Fix potential double deletion when killing entities
Fix potential double deletion when killing entities. If an entity
has already been marked as "killed" it may not be added to the set
of entities to be killed again since this would lead into double
deletion later on since the entity in question will be deleted
at the end of the current update loop.