Use a well defined order for persisting custom material properties
Use a well defined order for persisting custom material properties.
Use an alphabetically sorted order for both custom shader uniforms
and texture maps when persisting these properties. The reason being
that the order of the unordered_map is undefined (like the name says)
and this can result in a change in the content file when no actual
data has changed. This change however will create a change in a version
control system such as Git which is not only unnecessary but also
rather annoying.