Refactor audio buffer allocation and buffer's size semantics
Refactor audio buffer allocation and buffer's size semantics.
Refactor the audio buffer allocation to use an abstract BufferAllocator
interface. This lets the elements to remain agnostic regarding the
actual buffer creation and allocation scheme. This is needed in order
to be able to transparently use a buffer provided by the platform's
audio system such as Waveout or Pulseaudio.

Refactor the buffer's size semantics in order to keep the allocated
capacity and actual content size separate.

Related issue #60