Add functionality for polling the widget for some action in uikit
Add functionality for polling the widget for some action in uikit.
For example when an increment/decrement button on a spinbox widget
is held the widget can adjust its value continuously. When the
value changes there should be a value changed event. This can now
be achieved through the use of PollAction as long as the client
code call this repeatedly.