Refactor render tree JSON serialization away from the tree
Refactor render tree JSON serialization away from the tree.
Having the serialization built in the tree makes the tree
less generic and makes it depend on the data/ interfaces which
hampers reusability. After this dependency is broken the tree
is completely generic and can be moved into base/ for reuse
in the uikit.