Refactor uikit widget composition into non-intrusive render tree
Refactor uikit widget composition into non-intrusive render tree.
Use the generic render tree from base/ in order to define the
widget tree and widgets' parent-child relationships. This change
simplifies the widget implementations and requires no longer widgets
to have child widgets. This has the benefit that questions like
"should the widget pass a mouse event to its child widgets?" etc.
don't ever exist since the widgets don't have children ever.

This commit changes the visible tree structure in the in editor
widget also so that the (nullptr) root of the render tree is now
also visible in the widget tree and the form widget is no longer
mandatory root of the hiearchy inside the window. In fact it's now
possible to create windows with any number of form widgets
(including zero forms).