Refactor style loading in uiwidget in GSEditor
Refactor style loading in uiwidget in GSEditor. Make sure that there's
always a valid c++ style object. Refactor the style loading into
verbose/quiet functions. The verbose function can be used in an
interactive manner when the user for example clicks on style reload.
The quiet function is intended for when the style loading happens
on the background (i.e. on widget load).