Add window event logging and replay to playwindow in GSEditor
Add window event logging and replay to playwindow in GSEditor. This
will allow one to record the mouse/keyboard events in order to be
able to replay the same sequence of events at later time. The chief
motivation for doing is that when combined with some other utilities
and macro type functionality in playwindow it'd be possible to do
automated system testing. I.e. load up a scene/UI replay an event log
and then verify the expected output.

TODO: Recording the mouse move events is currently very primitive
and potentially records a lot of data. This will need some more
refinement in order to bring the volume of data down. Either by
having a way to change the capture resolution on mouse move or by
intelligently capturing mouse move events so that the same events
can then be synthesized on replay.

TODO: Maybe refactor away the event object interface and simply
use a std::variant.