Move looped audio playback implementation to lower layer
Move looped audio playback implementation to lower layer in the
audio stack. The problem with doing looped audio playback in the
audio player (or higher up in the stack) is that if a new device
PCM playback stream is created there might be gaps or overlaps in
the audible playback depending on the device, PCM buffer sizes,
device latency etc. since there are now infact 2 indepdent streams
that aren't synchronized. Using a single stream should eliminate
these problems. Additionally the meaning of "looping" might not be
meaningful to some sources such as sources that continuously generate
PCM data.