Refactor play window main loop and window event dispatching
Refactor play window main loop and window event dispatching. This
makes the play window's "main loop" *both* run atomically *and*
process events and perform engine calls exactly in the same order
as the real game main loop.

Previously the issue was that the window events were dispatched
from the play window's eventFilter callback directly which meant
that the Qt event loop dispatch was sandwitched between calls to
PlayWindow::BeginPlay and PlayWindow::Update. This however had the
unwanted side effect (bug) that because the window initialization
uses a timer the init state would happen between the BeginMainLoop
and Update calls thus violating the invariant that Update is never
called without the previous call to BeginMainLoop first.

This commit fixes all of that by introducing a queue (ring buffer)
in which the Qt's window events are enqueued until the main loop
function runs. It's then trivial to make sure that we make all the
right calls in the right order to BeginMainLoop, dispatch the window
events, Update, Draw and finally EndMainLoop.