Add some very primitive tracing utilities for "profiling" the engine
Add some very primitive tracing utilities for "profiling" the engine.
The problem is that tools like Valgrind (with callgrind) cannot be
used on a whole application since it's dog slow. Other tools such as
NVIDIA NSIGHT are not well supported on Linux. A simple trace of
top most engine functions with some timing information can already
provide some clues if/when frame spikes happen and where.