Limit the use of std::random_device to seeding in RandomString
Limit the use of std::random_device to seeding in RandomString
generation. The problem is that the std::random_device tries to
provide random numbers by sampling the entropy in the system and
can subsequently be very slow. This would not really matter for
generating class IDs but unfortunately there are some code paths
that call this function when creating instance IDs. Ultimately
these need to be refactored but for now refactor the RandomString
to use pseudo random generator only.