Refactor OpenGL state setting in device implementation
Refactor OpenGL state setting in device implementation. This change
moves all the state setting closer together which makes it easier
to reason about the code and understand all the state that is being
set and helps in avoid any such bugs. Additionally it'll be easier
to add caching to reduce redundant state calls.