Add flags to control garbage collection in gfx::Device
Add flags to control garbage collection in gfx::Device. Currently
the expectation is that during actual game play only textures that
are used to render text need to be reclaimed and cleaned away once
they're no longer being used. Other resources would always remain
in use and should not be deleted. For the editor however things
change since during the editing plenty of stuff can be be created
which is then later discarded and not useful.