Add graphics flag to indicate editing mode in the environment
Add graphics flag to indicate editing mode in the environment. The
fundamental problem is balancing between modifying content in the
editor and checking for changes at runtime. If some resource is marked
as "static" it means it'll not be changed during *game play*. It can
however  be changed during *editing*. This means that if we don't
check static content for changes the changes that are done while
editing will not be reflected. On the other hand if we do check for
changes also during game play then effort is wasted to check something
that should not change. The solution is to have an environment flag
that indicates whether the environment is static or not (editing mode
vs. game mode) and then choose the right action when checking for
changes in the resource data.

The place where to turn this flag on/off is the painter since all
high level drawing operations go through the painter which will then
already setup the environment parameters that are passed to both the
material and drawable objects when applying their state and uploading
the GPU resources.