Reuse a single polygon class to visualize the current draw command
Reuse a single polygon class to visualize the current draw command
when working on creating a new polygon/shape class in the editor.
This avoids creating needless geometry objects when one will suffice.

Likewise use a static color class for the material instead of
creating a new color class continuously.