Fix shader program uniform state setting bug in commit b4269787
Fix shader program uniform state setting bug in commit b4269787

The root bug is that setting the shader program state is skipped when
the geometry doesn't have any vertex data. This also leaves the program
state unmodified by skipping setting the program  uniforms. However
the cached uniform value is kept which on next draw leads to the
incorrect conclusion that there's no need to set program uniform since
the value has already been set (when it hasn't).

The simplest fix now (while keeping the uniform hashing and caching)
is to always set the uniforms state into the shader program on the
first draw when the program is used regardless of whether the draw
will do an early exit or not.