Add abstract material interface for material instances
Add abstract material interface for material instances. The idea
behind this interface is to separate any particular material instance
implementation from the interface. There are cases when it might be
desireable to have a simple material implementation without any
actual class object. For example when drawing transient text using
text rasterized in a bitmap. This will allow the text rendering to
be separated from the rest of the material system and hopefully allow
for more leeway in the future regarding how the dynamic text rendering
will be implemented.