Refactor displaying text and how it relates to the material system
Refactor displaying text and how it relates to the material system.
Currently we're using a texture to display a whole block of text
that has been prerasterized into a bitmap. The problem is that some
of the text can be "transient" in nature, i.e. only displayed for a
short time. This creates friction regarding how to do the texture
uploads in the material system since the "normal" use case is that
"normal" textures (sprite, texture maps) are rather static and not
changed dynamically and don't need to be ever cleaned up.
Checking for modifications when binding textures slows things down
when dealing with these static textures and the whole concept of
garbage collection gets messy since a special flag is needed to
differentiate between textures that should be collected away and
those that shouldn't (i.e. transient text vs. static content).
Additionally this garbage collection flag is not otherwise needed
at the material level so some additional baggage is added there.

This commit refactors the way text rendering works by removing the
complications mentioned above from the "normal" material texture
upload path. Instead a new texture instance class is provided with
the only intention being to be used for displaying text dynamically.
It uses the same idea of using scratch/transient textures that have
the whole text pre-rasterized in them. However this now provides a
completely alternate code path and thus simplifies the regular material
texture uploading/binding functionality and allows that to be optimally
efficient when using static textures/content. (No need to check for
modifications when everything is marked "static") TODO: The static
flags concerning textures should probably be moved to be specific to
a texture instead of being derived from the material's static flag.

The garbage collection flag is removed from the material level. Instead
there's now a transient flag that is only at the texture level and the
cleanup transient textures is now *different* from the garbage collection.
Garbage collection is now *only* for removing crap that has been
needlessly created (i.e. during editing of content in the GSEditor) and
is no longer needed and cleaning up transient textures used to display
text is now completely separate from that. This means that the actual
engine no longer needs to call CleanGarbage since there should not be
any GPU resources that are created without being needed.

The new TextMaterial fragment shader also now correctly samples the
texture for the alpha and uses that as the *alpha* value of the
gl_FragColor instead of multiplying the color value with the alpha.
This change causes all the text rendering related gold images to
change (for the better).