Set a font to first font available when enabling entity node text item
Set a font to first font available when enabling entity node text item
in the GSEditor. Without a font selected text cannot be rasterized.
This will lead to spammy logs when on each render the system will
try to rasterize the text and fails. Since it's difficult to provide
validation at the UI level for only this part of the entity rendering
(could probably be done through a render hook..) we're just going to
use some font setting as a default.