Add entity node text item raster buffer width and height
Add entity node text item raster buffer width and height. If not set
(i.e. 0) then the renderer will derive the size based on the node's
current size. However this will mean if the node's size changes
(because of changes in the render tree) the text item contents will
need to be rerasterized. This can cause performance penalty. Using
a fixed size will prohibit this and always only rasterize once.