Fix node's transformation reset on animation track restart
Fix node's transformation reset on animation track restart. When an
animation track is restarted the nodes' transformations are reset.
However the top level nodes (i.e. the entity's parent nodes, i.e. the
nodes that don't have parents, i.e. the nodes that are linked to the
scene's root) can't have their transformations reset to what is in
the entity class simply because they have their spawning placement
baked into their transforms and cannot thus be reset to whatever is
in the class nodes's transformation.