Add audio subsystem in the WASM build and enable in the test engine
Add audio subsystem in the WASM build and enable in the test engine.

Add a workaround for streams going into AL_STOPPED state randomly when
using the OpenAL implementation from Emscripten. Tested with Chromium
Version 96.0.4664.45. The workaround is simply to try to put the
stream back into play state with alSourcePlay.