Add feature flag checks to potentially enable SSE2 in WASM build
Add feature flag checks to potentially enable SSE2 in WASM build
when building libvorbis and libsndfile using emscripten. Enabling
the SSE2 instruction set then requires -sse2 -msimd128 compiler
flags to be turned on after which the SSE2 instructions will be used.

However, it seems that turning the SSE2 based float/double -> int
conversions in libsndfile breaks something and creates distortion
in the audio when decoding ogg/vorbis. (Works fine on the desktop).
This could potentially be an issue in the SSE2 implementation in
Emscripten (_mm_load_ss and _mm_load_sd are emulated) or an issue
related to alignment of the SIMD data.