Throw exceptions on Lua script errors when loading
Throw exceptions on Lua script errors when loading. The rationale
behind this change is that the Lua error is something we really
want to propagate all the way to the UI when running in the editor.
It's quite possible to have mistakes in the code and this should
be as loud and visible as possible. Throwing and exception is a
super simple way to achieve this.