Add support for packaging for Web using Emscripten tools
Add support for packaging for Web using Emscripten tools. This is
still slightly finicky and requires a) python b) emscripten sdk
and that the user configures these properly in the editor's tools.
Long term plan would be to package all the required stuff with the
Gamestudio itself in order to reduce the barrier to entry for
publishing to the web.