Fix incorrect application path (app://) resolution in engine loader
Fix incorrect application path (app://) resolution in engine loader.
This bug was introduced in commit c385685d which refactored the
game content packaging into a sub-folder in the output folder. This
change however now means that resources that are identified as app
resources (i.e. starting with app://) no longer resolve correctly
since the application resource path no longer matches with the
content package path. Before the change it matched through a sheer
coincidence so the underlying issue was masked.

It might make sense to package the resources that originally come
from the editor resources separately. If that is ever changed then
this URI resolution in the engine loader must also change.