Add a persistent IDBFS filesystem mounting and syncing to web main
Add a persistent IDBFS filesystem mounting and syncing to web main.
The problem is that the default filesystem in Emscripten is only
in memory and is not persistent anywhere. We can try to work around
that by using a IndexedDB based filesystem that is mounted in
addition to the default filesystem. However, the problem with this
filesystem is that the operations are async. To work around this
the main emscripten/wasm application tries to defer the actual
game loading until the filesystem mount has finished (indicated
via a JS callback). Likewise a flush/sync operation is attempted
when the game exits and the engine has performed orderly shutdown.

https://developer.mozilla.org/en-US/docs/Web/API/IndexedDB_API