Add application log event flags in the project settings
Add application log event flags in the project settings. These can
be used to provide some settings in an environment where it might
not be so simple to pass them at runtime. For example when running
the game in the browser enabling debug logs would require being able
to pass the debug flag as a query parameter and then generate the
html5  page so that parameter can then be passed to the WASM blob
on load before the actual application execution starts.