Add audio SourceThreadProxy wrapper for audio sources using a thread
Add audio SourceThreadProxy wrapper for audio sources using a thread.
The original intention was to able to use this with WASM since the
OpenAL audio implementation provided by Emscripten (and the underlying
browser's audio API) aren't made to be used from a background thread.
So instead of using a thread for the audio device a thread could be
used for the audio source evaluation. Of course the issue here is
that this still will require the right access policy before the
threading (through SharedArrayBuffer) works so in practice threading
insn't available in WASM.