Disable mipmaps and mipmapped texture filtering in gfx::TextMaterial
Disable mipmaps and mipmapped texture filtering in gfx::TextMaterial.
Fundamentally the assumption is that the text is being drawn in a
such a way that the rasterized geometry matches with the size of the
rasterized text texture/buffer in which case a simple fast nearest
texture filtering can be used and there's no need to generate mips
either, thus saving some useless work from being done.