Enable custom tracing in emscripten/WASM engine build
Enable custom tracing in emscripten/WASM engine build. For saving
the captured JSON trace file from the memory only file system in
the WASM environment a HTML UI is needed with a 3rd party JS file
for using some native browser functionality.

A copy of FileSaver.js is provided in the editor/dist folder but
is not deployed by default when a game package is built.