Add entity flag for controlling entity killing at the scene boundary
Add entity flag for controlling entity killing at the scene boundary.
The idea of this flag is that when an entity (and all of its entity
nodes) are outside of the specified scene boundary it'll be killed
automatically by the scene update. This will simplify things such
defining a border when a player character that is falling downwards
is killed automatically or killing objects that are thrown or moved
out of the relevant scene game play area.