Use automatic entity killing at scene boundary in Blast demo
Use automatic entity killing at scene boundary in Blast demo
except with the enemy ships. They're spawned at a position that
is *intentionally* outside of the scene and would then get killed
immediately. Leaving this as-is for now since it's unclear whether
more granular kill flag setting would make sense or not. Possibly
letting the game to deal with this edge case is totally fine.