Add an editor setting to change the default gfx widget clear color
Add an editor setting to change the default gfx widget clear color.
Refactor the hotkey to invoke the clear color dialog to using a
Shift modifier key. The reason being that we also need a hotkey
for resetting the widget specific clear color so that the generic
setting will take effect and this second hotkey would collide with
the VSYNC toggle hotkey and other simple function key based hotkeys.