Set cursor position to 0 when setting text on QLineEdit
Set cursor position to 0 when setting text on QLineEdit. The default
is to have the cursor at the end of the text which often results in
only the tail of the text being visible in the QLineEdit widget
and in cases where the whole text doesn't fit showing the start
instead of the end of the text seems like a better choice.