Cache designated FileSource PCM buffers after complete decode
Cache designated FileSource PCM buffers after complete decode. Some
audio sounds that are played frequently and/or a rapid succession
(such as the player's weapon sound in Blast demo) can be cached
in order to reduce the repeated work that is done to load the
contents and decode the audio files. Caching the raw PCM data can
improve performance by a factor of 10 in the perf test application.
The downside of course is that this will require more RAM usage
since the raw PCM data can be quite large. It's therefore intended
that only very few select audio files would ever have the PCM
caching flag set.