Refactor audio stream loading from std::ifstream to SourceStream
Refactor audio stream loading from std::ifstream to SourceStream.
Using a more elaborate interface should allow better ways to create
several alternative implementations for stream based audio loading.

The yet unresolved design level problem however is whether there
should exists multiple interfaces/ways to load audio data in the
first place. Semantically there's currently a small difference in
the sense that the buffers are considered to be immutable blobs of
data that can be shared between multiple audio objects for reading
whereas the streams have state and cannot be shared.

The dependant design issue is that because of these two different
ways to access the audio data there are multiple IO device decoder
implementations for both SndFile and Mpg123. If there was just one
way to load the data these abstractions could be removed.

Experience, however, tells me that these types of designs tend to
go in circles so I'm leaving these interfaces as-is for now.