Refactor the physics subsystem in engine
Refactor the physics subsystem in engine to use different functions
for different purpose. Updating the physics world based on the
changes in rigid bodies (such as flags or static body position) is
now done in one function and updating the scene based on the physics
simulation is done in another function. Splitting the functionality
into separate functions simplifies the code and makes it clearer
how data is reflected from scene to physics and back.