Fix incorrect tests in base/unit_test.cpp
Fix incorrect tests in base/unit_test.cpp. Commit e0a55fe1a86
changed the semantics of rectangle border values to be inclusive.
This broke these base test but that was not discovered until now
because the tests were not being run because of accidental omission
of add_test in the project CMakeLists.