Add entity method to let the entity to die and be removed from the scene
Add entity method to let the entity to die and be removed from the scene.
We set a flag to indicate that the entity has expressed its wish
to die and on scene loop end move the entity to the killset for
the normal kill state transitions and later deletion.