Let Lua errors propagate out of any call into Lua game code
Let Lua errors propagate out of any call into Lua game code. When
the Lua code has faulted the best thing to do is to represent the
error and the Lua game code stack trace to the game programmer. To
that extent it's better to propagate this information to the caller
instead of simply logging it.

Reword and reformulate the rationale and possible alternatives and
strategies regarind how to handle both engine and Lua game errors.