Relayout the entity widget UI and remove widget scroll area
Relayout the entity widget UI and remove widget scroll area. The
scroll area makes it difficult to control the layout and not have
scrollbars appear when they're not wanted. Since the the entity node
attachments/items are in a TabWidget already there's no need for
vertical scrolling all that much anyway.