Add support for attaching additional fixtures to rigid bodies
Add support for attaching additional fixtures to rigid bodies in
the physics subsystem. When a node with a fixture attachment is
found the attachement is used to create an additional physics fixture
(collision shape) and attach it to an existing physics body. The
current limitation is that the rigid body must have been created
before the fixture creation is attempted. In other words the
entity node that has the rigid body which the fixture is attached
must be a parent node of the node with the fixture attachment. When
this is true the parent node is visited first and the rigid body
is created first before the fixture.

This commit refactors the implementation of the physics subsystem.

- PhysicsNode is split into 2 meta data structures:
  * RigidBodyData that has meta data relevant to a rigid body only.
  * FixtureData that has meta data relevant to a fixture.
- Debug drawing must now traverse all the fixtures attached to a body
  and draw "fixtures" instead of "nodes".
- Collision shape vertices must be transformed including a possible
  rotation so that collision shape can be rotated arbitrarily relative
  to the physics body local coordinates.

TODO:
- Remove the limitation regarding the entity node creation order,
  rigid body and fixture creation order.