Add texture tag for texture grouping and group identification
Add texture tag for texture grouping and group identification.
Currently the device garbage collection looks at the frame stamps
to determine which objects should be reclaimed. This has the problem
that we might accidentally clean away textures that are part of a
 sprite cycle but have not been recently in the current phase of the
cycle. With these sprite batches a preferred way would be to treat
them as a group for cleaning.

The group tag will now now be used in the garbage collection cycle
to determine whether all textures in a group (sprite batch) have
expired or not. Cleaning up then only happens when all the textures
in the group have expired.