Fix repeated texture uploads in DlgBitmap and DlgText
Fix repeated texture uploads in DlgBitmap and DlgText. Creating the
gfx resources on the stack generates new unique IDs for the materials
and this causes content to be uploaded to the GPU on every paint.
Fix this by creating the gfx resources only one and then update
their content properties instead of creating new objects.