Fix incorrect texture cleanup when shader doesn't use the uniform
Fix incorrect texture cleanup when shader doesn't use the uniform.
THe problem is that if the texture is not actually used by the shader
the shader compiler can optimize it away. When this happens the draw
implementation doesn't actually use the texture and ends up skipping
the frame stamping. When the frame stamp is not updated the texture
gets cleaned away. The problem with this is that the material system
doesn't know about the fact that the texture isn't actually used and
then ends up creating a new texture object and uploads it again since
the previous instance was deleted.

Fix the above issue by always updating the texture frame stamp and
only then skip the texture binding if the uniform isn't actually
available in the program.